+++ src/main/java/com/havis/object/post/model/dto/PostListResponseDTO.java
@@ -1,5 +1,6 @@
-package com.havis.object.post.model.entity;
+package com.havis.object.post.model.dto;
 
+import com.havis.object.post.model.entity.PostEntity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -23,13 +24,13 @@ public class PostListResponseDTO {
     private LocalDateTime modifiedAt;
 
     // Entity -> dto
-    public PostListResponseDto(PostEntity postEntity) {
+    public PostListResponseDTO(PostEntity postEntity) {
         this.postTitle = postEntity.getPostTitle();
         this.createdAt = postEntity.getModifiedAt();
         this.modifiedAt = postEntity.getCreatedAt();
     }
 
-    public PostListResponseDto(Optional<PostEntity> postEntity) {
+    public PostListResponseDTO(Optional<PostEntity> postEntity) {
         this.postTitle = postEntity.get().getPostTitle();
         this.createdAt = postEntity.get().getModifiedAt();
         this.modifiedAt = postEntity.get().getCreatedAt();
