+++ src/main/java/com/havis/object/post/model/dto/PostRequestDTO.java
@@ -1,4 +1,4 @@
-package com.havis.object.post.model.entity;
+package com.havis.object.post.model.dto;
 
 import com.havis.object.category.model.entity.CategoryEntity;
 import com.havis.object.member.model.entity.MemberEntity;
