+++ src/main/java/com/havis/object/post/repository/PostRepository.java
@@ -1,5 +1,7 @@
-package com.havis.object.post.model.entity;
+package com.havis.object.post.repository;
 
+import com.havis.object.post.model.dto.PostListResponseDTO;
+import com.havis.object.post.model.entity.PostEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
